Fix conversion from/to fty-proto
