+++ src/fty_asset_manipulation.cc
@@ -35,7 +35,9 @@
 
 #include <tntdb/row.h>
 #include <tntdb/connect.h>
-#include <fty_common_asset.h>
+
+#include <ftyproto.h>
+#include <zhash.h>
 
 /// create asset name from type/subtype and integer ID
 static std::string createAssetName(const std::string& type, const std::string& subtype, long index)
@@ -218,4 +220,90 @@ fty::Asset updateAsset(const fty::Asset& asset, bool test)
     }
 
     return updatedAsset;
+}
+
+// for fty-proto conversion helpers
+static fty::Asset::ExtMap zhashToExtMap(zhash_t *hash, bool readOnly)
+{
+    fty::Asset::ExtMap map;
+
+    for (auto* item = zhash_first(hash); item; item = zhash_next(hash))
+    {
+        map.emplace(zhash_cursor(hash), std::make_pair(static_cast<const char *>(item), readOnly));
+    }
+
+    return map;
+}
+
+
+static zhash_t* extMapToZhash(const fty::Asset::ExtMap &map)
+{
+    zhash_t *hash = zhash_new ();
+    for (const auto& i :map) {
+        zhash_insert (hash, i.first.c_str (), const_cast<void *>(reinterpret_cast<const void *>(i.second.first.c_str())));
+    }
+
+    return hash;
+}
+
+// fty-proto/Asset conversion
+fty_proto_t * assetToFtyProto(const fty::Asset& asset, const std::string& operation, bool test)
+{
+    fty_proto_t *proto = fty_proto_new(FTY_PROTO_ASSET);
+
+    zhash_t *aux = zhash_new();
+    zhash_autofree (aux);
+
+    zhash_insert(aux, "priority", const_cast<void *>(reinterpret_cast<const void *>(std::to_string(asset.getPriority()).c_str())));
+    zhash_insert(aux, "type", const_cast<void *>(reinterpret_cast<const void *>(asset.getAssetType().c_str())));
+    zhash_insert(aux, "subtype", const_cast<void *>(reinterpret_cast<const void *>(asset.getAssetSubtype().c_str())));
+    if (test)
+    {
+        zhash_insert(aux, "parent", const_cast<void *>(reinterpret_cast<const void *>(0)));
+    }
+    else
+    {
+        zhash_insert(aux, "parent", const_cast<void *>(reinterpret_cast<const void *>(selectAssetProperty<int>("id_parent", "name", asset.getInternalName()))));
+    }
+    zhash_insert(aux, "status", const_cast<void *>(reinterpret_cast<const void *>(assetStatusToString(asset.getAssetStatus()).c_str())));
+
+    zhash_t *ext = extMapToZhash(asset.getExt());
+
+    fty_proto_set_aux(proto, &aux);
+    fty_proto_set_name(proto, "%s", asset.getInternalName().c_str());
+    fty_proto_set_operation(proto, "%s", operation.c_str());
+    fty_proto_set_ext(proto, &ext);
+
+    zhash_destroy(&aux);
+    zhash_destroy(&ext);
+
+    return proto;
+}
+
+fty::Asset ftyProtoToAsset(fty_proto_t * proto, bool extAttributeReadOnly, bool test)
+{
+    if (fty_proto_id(proto) != FTY_PROTO_ASSET)
+    {
+        throw std::invalid_argument("Wrong fty-proto type");
+    }
+
+    fty::Asset asset;
+    asset.setInternalName(fty_proto_name(proto));
+    asset.setAssetStatus(fty::stringToAssetStatus(fty_proto_aux_string(proto, "status", "active")));
+    asset.setAssetType(fty_proto_aux_string(proto, "type", ""));
+    asset.setAssetSubtype(fty_proto_aux_string(proto, "subtype", ""));
+    if(test)
+    {
+        asset.setParentIname("");
+    }
+    else
+    {
+        asset.setParentIname(selectAssetProperty<std::string>("name", "id_parent", fty_proto_aux_number(proto, "parent", 0)));
+    }
+    asset.setPriority(fty_proto_aux_number(proto, "priority", 5));
+
+    zhash_t *ext = fty_proto_ext(proto);
+    asset.setExt(zhashToExtMap(ext, extAttributeReadOnly));
+
+    return asset;
 }
